consolidating the call to /Card component for all pages
